pyproject.toml
[tool.poetry]
name = "ai-favs"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.23.2"
pydantic = "<2"
python-dotenv = "^1.0.1"
openai = "^1.17.1"
faiss-cpu = "^1.8.0"
langchain-cli = ">=0.0.15"
langchain-openai = "^0.1.3"
langserve = {extras = ["server"], version = ">=0.0.30"}
langchain-community = "^0.0.36"
langsmith = "^0.1.51"
langchain-text-splitters = "^0.0.1"
langchain-core = "^0.1.50"
langchain = "^0.1.17"
playwright = "^1.43.0"
sqlalchemy = "^2.0.30"
pydub = "^0.25.1"
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
